+++ editor/src/main/java/com/canopas/editor/ui/RichTextStyle.kt
@@ -1,6 +1,7 @@
 package com.canopas.editor.ui
 
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.ParagraphStyle
 import androidx.compose.ui.text.PlatformSpanStyle
 import androidx.compose.ui.text.SpanStyle
@@ -10,12 +11,11 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.font.Typeface
 import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.unit.TextUnit
+import androidx.compose.ui.unit.em
 import androidx.compose.ui.unit.sp
 
 interface RichTextStyle {
     fun applyStyle(spanStyle: SpanStyle): SpanStyle
-    fun applyStyle(paragraphStyle: ParagraphStyle): ParagraphStyle
-
     fun isTitleStyles(): Boolean
 
     object Bold : RichTextStyle {
@@ -24,22 +24,29 @@ interface RichTextStyle {
 
         }
 
-        override fun applyStyle(paragraphStyle: ParagraphStyle): ParagraphStyle {
-           return paragraphStyle
-        }
-
         override fun isTitleStyles() = false
     }
 
     object TITLE : RichTextStyle {
         override fun applyStyle(spanStyle: SpanStyle): SpanStyle {
             return spanStyle.copy(
-                fontSize = 44.sp,
-                fontWeight = FontWeight.Medium
+                fontSize = 38.sp,
+                fontWeight = FontWeight.Medium,
+                letterSpacing = 0.15.sp
             )
         }
-        override fun applyStyle(paragraphStyle: ParagraphStyle): ParagraphStyle {
-            return paragraphStyle
+
+        override fun isTitleStyles() = true
+    }
+
+    object SUB_TITLE : RichTextStyle {
+        override fun applyStyle(spanStyle: SpanStyle): SpanStyle {
+            return spanStyle.copy(
+                fontSize = 15.sp,
+                fontWeight = FontWeight.Medium,
+                color = Color(0XFF666666),
+                letterSpacing = 0.15.sp
+            )
         }
 
         override fun isTitleStyles() = true
@@ -53,9 +60,7 @@ interface RichTextStyle {
                 fontWeight = FontWeight.Medium
             )
         }
-        override fun applyStyle(paragraphStyle: ParagraphStyle): ParagraphStyle {
-            return paragraphStyle
-        }
+
         override fun isTitleStyles() = true
 
     }
@@ -63,44 +68,34 @@ interface RichTextStyle {
     object H2 : RichTextStyle {
         override fun applyStyle(spanStyle: SpanStyle): SpanStyle {
             return spanStyle.copy(
+                fontSize = 28.sp,
+                fontWeight = FontWeight.Medium
+            )
+        }
+
+        override fun isTitleStyles() = true
+    }
+
+    object H3 : RichTextStyle {
+        override fun applyStyle(spanStyle: SpanStyle): SpanStyle {
+            return spanStyle.copy(
                 fontSize = 24.sp,
                 fontWeight = FontWeight.Medium
             )
         }
-        override fun applyStyle(paragraphStyle: ParagraphStyle): ParagraphStyle {
-            return paragraphStyle
-        }
-        override fun isTitleStyles() = true
 
+        override fun isTitleStyles() = true
     }
 
-    object H3 : RichTextStyle {
+    object H4 : RichTextStyle {
         override fun applyStyle(spanStyle: SpanStyle): SpanStyle {
             return spanStyle.copy(
                 fontSize = 20.sp,
                 fontWeight = FontWeight.Medium
             )
         }
-        override fun applyStyle(paragraphStyle: ParagraphStyle): ParagraphStyle {
-            return paragraphStyle
-        }
+
         override fun isTitleStyles() = true
-
-
-    }
-
-    object H4 : RichTextStyle {
-        override fun applyStyle(spanStyle: SpanStyle): SpanStyle {
-            return spanStyle.copy(
-                fontSize = 18.sp,
-                fontWeight = FontWeight.Medium
-            )
-        }
-        override fun applyStyle(paragraphStyle: ParagraphStyle): ParagraphStyle {
-            return paragraphStyle
-        }
-        override fun isTitleStyles() = true
-
     }
 
     object H5 : RichTextStyle {
@@ -110,11 +105,8 @@ interface RichTextStyle {
                 fontWeight = FontWeight.Medium
             )
         }
-        override fun applyStyle(paragraphStyle: ParagraphStyle): ParagraphStyle {
-            return paragraphStyle
-        }
-        override fun isTitleStyles() = true
 
+        override fun isTitleStyles() = true
     }
 
 
@@ -125,11 +117,8 @@ interface RichTextStyle {
                 fontWeight = FontWeight.Medium
             )
         }
-        override fun applyStyle(paragraphStyle: ParagraphStyle): ParagraphStyle {
-            return paragraphStyle
-        }
-        override fun isTitleStyles() = true
 
+        override fun isTitleStyles() = true
     }
 
 
@@ -137,9 +126,7 @@ interface RichTextStyle {
         override fun applyStyle(spanStyle: SpanStyle): SpanStyle {
             return spanStyle.copy(fontStyle = FontStyle.Italic)
         }
-        override fun applyStyle(paragraphStyle: ParagraphStyle): ParagraphStyle {
-            return paragraphStyle
-        }
+
         override fun isTitleStyles() = false
     }
 
@@ -160,11 +147,8 @@ interface RichTextStyle {
                 }
             )
         }
-        override fun applyStyle(paragraphStyle: ParagraphStyle): ParagraphStyle {
-            return paragraphStyle
-        }
-        override fun isTitleStyles() = false
 
+        override fun isTitleStyles() = false
     }
 
     data class FontSize(val fontSize: TextUnit) : RichTextStyle {
@@ -175,10 +159,7 @@ interface RichTextStyle {
                 fontWeight = FontWeight.Normal
             )
         }
-        override fun applyStyle(paragraphStyle: ParagraphStyle): ParagraphStyle {
-            return paragraphStyle
-        }
-        override fun isTitleStyles() = false
 
+        override fun isTitleStyles() = false
     }
 }
